all constraints + one requestEco + more refactoring
