+++ Server/Util.cs
@@ -1,0 +1,86 @@
+﻿using Server;
+using System;
+using System.IO;
+using System.Net.Sockets;
+using System.Text;
+using System.Text.Json;
+
+namespace Server2
+{
+    public static class Util
+    {
+        public static Response ReadResponse(this TcpClient client)
+        {
+            var strm = client.GetStream();
+            byte[] resp = new byte[2048];
+            using (var memStream = new MemoryStream())
+            {
+                int bytesread = 0;
+                do
+                {
+                    bytesread = strm.Read(resp, 0, resp.Length);
+                    memStream.Write(resp, 0, bytesread);
+                    var responseData2 = Encoding.UTF8.GetString(memStream.ToArray());
+                } while (bytesread == 2048);
+
+                var responseData = Encoding.UTF8.GetString(memStream.ToArray());
+                return JsonSerializer.Deserialize<Response>(responseData, new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase });
+            }
+        }
+        public static Request ReadRequest(this TcpClient client)
+        {
+            var strm = client.GetStream();
+            //strm.ReadTimeout = 250;
+            byte[] resp = new byte[2048];
+            using (var memStream = new MemoryStream())
+            {
+                int bytesread = 0;
+                do
+                {
+                    bytesread = strm.Read(resp, 0, resp.Length);
+                    memStream.Write(resp, 0, bytesread);
+                    var responseData2 = Encoding.UTF8.GetString(memStream.ToArray());
+                } while (bytesread == 2048);
+
+                var requestData = Encoding.UTF8.GetString(memStream.ToArray());
+
+
+                var test = JsonSerializer.Deserialize<Request>(requestData, new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase });
+                Console.WriteLine(test);
+                return JsonSerializer.Deserialize<Request>(requestData, new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase });
+            }
+        }
+
+        public static Boolean ArrayContains(string[] array, string stringToCheck)
+        {
+            if (stringToCheck == null)
+            {
+                return false;
+            }
+            foreach (string x in array)
+            {
+                if (stringToCheck.Contains(x))
+                {
+                    return true;
+                }
+            }
+            return false;
+        }
+        public static string ToJson(this object data)
+        {
+            return JsonSerializer.Serialize(data, new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase });
+        }
+        public static T FromJson<T>(this string element)
+        {
+            return JsonSerializer.Deserialize<T>(element, new JsonSerializerOptions { PropertyNamingPolicy = JsonNamingPolicy.CamelCase });
+        }
+        public static DateTime UnixTimeStampToDateTime(int unixTimeStamp)
+        {
+            // Unix timestamp is seconds past epoch
+            System.DateTime dtDateTime = new DateTime(1970, 1, 1, 0, 0, 0, 0, System.DateTimeKind.Utc);
+            dtDateTime = dtDateTime.AddSeconds(unixTimeStamp).ToLocalTime();
+            return dtDateTime;
+        }
+
+    }
+}
